requirements.txt
streamlit>=1.28.2
requests>=2.31.0
openai>=1.3.5
pydantic>=2.5.2
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
gunicorn
supabase>=2.0.3
beautifulsoup4>=4.12.2
python-multipart>=0.0.6
